Add a failing cli test
This change create a complete end to end test showing a full example if
it the application is working correctly.

As you can see there is a concept of a somesthing called runtime. I am
a big believer in only testing my own code, as this example builds out
you will see that abstraction will make the code more testable.